# openapi.yaml
openapi: 3.0.3
info:
  title: Go Commerce API
  version: 1.0.0
  description: |
    REST API for the current Go Commerce implementation.
    The spec mirrors the running Gin gateway instead of describing future capabilities.
servers:
  - url: http://localhost:8080
...
paths:
...
  /api/register:
    post:
      tags: [Auth]
      summary: Register user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
      responses:
        '200':
          description: Registered
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /api/login:
    post:
      tags: [Auth]
      summary: Login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Logged in
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /api/products:
    get:
      tags: [Product]
      summary: List products
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/PageSize'
        - name: category
          in: query
          schema:
            type: string
        - name: keyword
          in: query
          schema:
            type: string
        - name: sort_by
          in: query
          schema:
            type: string
            enum: [created_at, price_cents, stock]
            default: created_at
        - name: order
          in: query
          schema:
            type: string
            enum: [asc, desc]
            default: desc
        - name: min_price_cents
          in: query
          schema:
            type: integer
            format: int64
            minimum: 0
        - name: max_price_cents
          in: query
          schema:
            type: integer
            format: int64
            minimum: 0
      responses:
        '200':
          description: Product page
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductListResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
  /api/products/{id}:
    get:
      tags: [Product]
      summary: Get product
      parameters:
        - $ref: '#/components/parameters/ID'
      responses:
        '200':
          description: Product
          content:
            application/json:
              schema:
                type: object
                properties:
                  product:
                    $ref: '#/components/schemas/Product'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /api/merchants:
    get:
      tags: [Merchant]
      summary: List merchants
      description: Current gateway implementation always requests page=1 and page_size=10.
      responses:
        '200':
          description: Merchant list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MerchantListResponse'
    post:
      tags: [Merchant]
      summary: Create merchant
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateMerchantRequest'
      responses:
        '200':
          description: Created
          content:
            application/json:
              schema:
                type: object
                properties:
                  merchant:
                    $ref: '#/components/schemas/Merchant'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /api/merchants/{id}:
    get:
      tags: [Merchant]
      summary: Get merchant
      parameters:
        - $ref: '#/components/parameters/ID'
      responses:
        '200':
          description: Merchant
          content:
            application/json:
              schema:
                type: object
                properties:
                  merchant:
                    $ref: '#/components/schemas/Merchant'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /api/merchants/products:
    post:
      tags: [Merchant]
      summary: Add product to explicit merchant
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AddMerchantProductRequest'
      responses:
        '200':
          description: Product created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AddProductResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
    delete:
      tags: [Merchant]
      summary: Delete product from explicit merchant
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DeleteMerchantProductRequest'
      responses:
        '200':
          description: Deleted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
  /api/merchant/profile:
    get:
      tags: [Merchant Console]
      summary: Get current merchant profile
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/MerchantID'
      responses:
        '200':
          description: Current merchant
          content:
            application/json:
              schema:
                type: object
                properties:
                  merchant:
                    $ref: '#/components/schemas/Merchant'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /api/merchant/products:
    get:
      tags: [Merchant Console]
      summary: List current merchant products
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/PageSize'
        - $ref: '#/components/parameters/MerchantID'
      responses:
        '200':
          description: Merchant products
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductListResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
    post:
      tags: [Merchant Console]
      summary: Create product for current merchant
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/MerchantID'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateConsoleProductRequest'
      responses:
        '200':
          description: Product created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AddProductResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /api/merchant/products/{id}:
    put:
      tags: [Merchant Console]
      summary: Update current merchant product
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/ID'
        - $ref: '#/components/parameters/MerchantID'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateConsoleProductRequest'
      responses:
        '200':
          description: Product updated
          content:
            application/json:
              schema:
                type: object
                properties:
                  product:
                    $ref: '#/components/schemas/Product'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
    delete:
      tags: [Merchant Console]
      summary: Delete current merchant product
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/ID'
        - $ref: '#/components/parameters/MerchantID'
      responses:
        '200':
          description: Deleted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /api/merchant/orders:
    get:
      tags: [Merchant Console]
      summary: List merchant-visible orders
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/PageSize'
        - $ref: '#/components/parameters/MerchantID'
      responses:
        '200':
          description: Merchant order page
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderListResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /api/orders:
    get:
      tags: [Order]
      summary: List current user's orders
      description: Current gateway implementation does not expose pagination parameters.
      security:
        - bearerAuth: []
      responses:
        '200':
          description: User orders
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderListResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalServerError'
    post:
      tags: [Order]
      summary: Create order
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateOrderRequest'
      responses:
        '200':
          description: Created order
          content:
            application/json:
              schema:
                type: object
                properties:
                  order:
                    $ref: '#/components/schemas/Order'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
  /api/orders/{id}:
    get:
      tags: [Order]
      summary: Get current user's order
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/ID'
      responses:
        '200':
          description: Order
          content:
            application/json:
              schema:
                type: object
                properties:
                  order:
                    $ref: '#/components/schemas/Order'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /api/orders/{id}/cancel:
    put:
      tags: [Order]
      summary: Cancel pending order
      description: Requires Idempotency-Key. Repeating the same key for the same user and order replays the first cancel response; reusing the key with a different order returns 409.
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/ID'
        - $ref: '#/components/parameters/IdempotencyKey'
      responses:
        '200':
          description: Cancel result
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '409':
          description: Idempotency key is already processing or was reused with a different cancel request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /api/orders/{id}/ship:
    put:
      tags: [Order]
      summary: Ship paid order
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/ID'
      responses:
        '200':
          description: Ship result
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
        '403':
          $ref: '#/components/responses/Forbidden'
        '409':
          $ref: '#/components/responses/Conflict'
  /api/orders/{id}/complete:
    put:
      tags: [Order]
      summary: Complete shipped order
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/ID'
      responses:
        '200':
          description: Complete result
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /api/payments:
    post:
      tags: [Payment]
      summary: Create payment
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreatePaymentRequest'
      responses:
        '200':
          description: Payment
          content:
            application/json:
              schema:
                type: object
                properties:
                  payment:
                    $ref: '#/components/schemas/Payment'
        '409':
          $ref: '#/components/responses/Conflict'
  /api/payments/{id}:
    get:
      tags: [Payment]
      summary: Get payment
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/ID'
      responses:
        '200':
          description: Payment
          content:
            application/json:
              schema:
                type: object
                properties:
                  payment:
                    $ref: '#/components/schemas/Payment'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /api/payments/{id}/success:
    post:
      tags: [Payment]
      summary: Mark payment succeeded
      description: Requires Idempotency-Key. Repeating the same key for the same user and payment replays the first success response; reusing the key with a different payment returns 409.
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/ID'
        - $ref: '#/components/parameters/IdempotencyKey'
      responses:
        '200':
          description: Payment updated
          content:
            application/json:
              schema:
                type: object
                properties:
                  payment:
                    $ref: '#/components/schemas/Payment'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          description: Payment is not actionable, the idempotency key is already processing, or the key was reused with a different payment success request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /api/payments/{id}/fail:
    post:
      tags: [Payment]
      summary: Mark payment failed
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/ID'
      responses:
        '200':
          description: Payment updated
          content:
            application/json:
              schema:
                type: object
                properties:
                  payment:
                    $ref: '#/components/schemas/Payment'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /api/cart:
    get:
      tags: [Cart]
      summary: Get cart
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Cart
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CartResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalServerError'
    delete:
      tags: [Cart]
      summary: Clear cart
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Cleared
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /api/cart/items:
    post:
      tags: [Cart]
      summary: Add cart item
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CartItemWriteRequest'
      responses:
        '200':
          description: Added
          content:
            application/json:
              schema:
                type: object
                properties:
                  item:
                    $ref: '#/components/schemas/CartItem'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
    put:
      tags: [Cart]
      summary: Update cart item
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CartItemWriteRequest'
      responses:
        '200':
          description: Updated
          content:
            application/json:
              schema:
                type: object
                properties:
                  item:
                    $ref: '#/components/schemas/CartItem'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
    delete:
      tags: [Cart]
      summary: Delete cart item
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [product_id]
              properties:
                product_id:
                  type: integer
                  format: int64
      responses:
        '200':
          description: Deleted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalServerError'
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    ID:
      name: id
      in: path
      required: true
      schema:
        type: integer
        format: int64
    Page:
      name: page
      in: query
      schema:
        type: integer
        minimum: 1
        default: 1
    PageSize:
      name: page_size
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 10
    MerchantID:
      name: merchant_id
      in: query
      schema:
        type: integer
        format: int64
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      description: Required idempotency key for write operations. Reuse with the same request replays the first response; reuse with a different request returns 409.
      schema:
        type: string
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Forbidden:
      description: Forbidden
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    NotFound:
      description: Not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Conflict:
      description: Conflict
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    InternalServerError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
  schemas:
    RegisterRequest:
      type: object
      required: [username, password, email]
      properties:
        username:
          type: string
        password:
          type: string
        email:
          type: string
        role:
          type: string
          enum: [customer, merchant]
    LoginRequest:
      type: object
      required: [username, password]
      properties:
        username:
          type: string
        password:
          type: string
    AuthResponse:
      type: object
      properties:
        user_id:
          type: integer
          format: int64
        token:
          type: string
        role:
          type: string
    Product:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        description:
          type: string
        price_cents:
          type: integer
          format: int64
          minimum: 0
        stock:
          type: integer
          format: int32
        category:
          type: string
        image_url:
          type: string
        merchant_id:
          type: integer
          format: int64
    ProductListResponse:
      type: object
      properties:
        products:
          type: array
          items:
            $ref: '#/components/schemas/Product'
        total:
          type: integer
          format: int64
    Merchant:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        contact_info:
          type: string
        created_at:
          type: string
          format: date-time
        owner_user_id:
          type: integer
          format: int64
    MerchantListResponse:
      type: object
      properties:
        merchants:
          type: array
          items:
            $ref: '#/components/schemas/Merchant'
        total:
          type: integer
          format: int64
    CreateMerchantRequest:
      type: object
      required: [name, contact_info]
      properties:
        name:
          type: string
        contact_info:
          type: string
    AddMerchantProductRequest:
      allOf:
        - $ref: '#/components/schemas/CreateConsoleProductRequest'
        - type: object
          required: [merchant_id]
          properties:
            merchant_id:
              type: integer
              format: int64
    DeleteMerchantProductRequest:
      type: object
      required: [merchant_id, product_id]
      properties:
        merchant_id:
          type: integer
          format: int64
        product_id:
          type: integer
          format: int64
    CreateConsoleProductRequest:
      type: object
      required: [name, description, price_cents, stock, category]
      properties:
        name:
          type: string
        description:
          type: string
        price_cents:
          type: integer
          format: int64
          minimum: 0
        stock:
          type: integer
          format: int32
        category:
          type: string
        image_url:
          type: string
    UpdateConsoleProductRequest:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        price_cents:
          type: integer
          format: int64
          minimum: 0
        stock:
          type: integer
          format: int32
        category:
          type: string
        image_url:
          type: string
    AddProductResponse:
      type: object
      properties:
        product_id:
          type: integer
          format: int64
    CreateOrderRequest:
      type: object
      required: [items]
      properties:
        items:
          type: array
          minItems: 1
          items:
            $ref: '#/components/schemas/CreateOrderItem'
    CreateOrderItem:
      type: object
      required: [product_id, quantity]
      properties:
        product_id:
          type: integer
          format: int64
        quantity:
          type: integer
          format: int32
    Order:
      type: object
      properties:
        id:
          type: integer
          format: int64
        user_id:
          type: integer
          format: int64
        items:
          type: array
          items:
            $ref: '#/components/schemas/OrderItem'
        total_amount_cents:
          type: integer
          format: int64
          minimum: 0
        status:
          type: string
          enum: [pending, paid, shipped, completed, cancelled]
        created_at:
          type: string
          format: date-time
        cancel_reason:
          type: string
          enum: ['', user_cancelled, payment_timeout]
    OrderItem:
      type: object
      properties:
        product_id:
          type: integer
          format: int64
        product_name:
          type: string
        price_cents:
          type: integer
          format: int64
          minimum: 0
        quantity:
          type: integer
          format: int32
    OrderListResponse:
      type: object
      properties:
        orders:
          type: array
          items:
            $ref: '#/components/schemas/Order'
        total:
          type: integer
          format: int64
    CreatePaymentRequest:
      type: object
      required: [order_id, payment_method]
      properties:
        order_id:
          type: integer
          format: int64
        payment_method:
          type: string
          enum: [mock_balance, mock_wechat, mock_alipay]
    Payment:
      type: object
      properties:
        id:
          type: integer
          format: int64
        payment_no:
          type: string
        order_id:
          type: integer
          format: int64
        user_id:
          type: integer
          format: int64
        amount_cents:
          type: integer
          format: int64
          minimum: 0
        status:
          type: string
          enum: [created, succeeded, failed]
        payment_method:
          type: string
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    CartItemWriteRequest:
      type: object
      required: [product_id, quantity]
      properties:
        product_id:
          type: integer
          format: int64
        quantity:
          type: integer
          format: int32
    CartItem:
      type: object
      properties:
        product_id:
          type: integer
          format: int64
        product_name:
          type: string
        price_cents:
          type: integer
          format: int64
          minimum: 0
        quantity:
          type: integer
          format: int32
        image_url:
          type: string
    CartResponse:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/CartItem'
        total_amount_cents:
          type: integer
          format: int64
          minimum: 0
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
    MessageResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
    ErrorResponse:
      type: object
      properties:
        error:
          type: string